[AIP-21] Move contrib.operators.docker_swarm_operator providers.docker.operators.docker_swarm
