+++ tests/providers/docker/__init__.py
@@ -16,36 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-"""
-from datetime import timedelta
-from airflow.utils.dates import days_ago
-from airflow import DAG
-from airflow.providers.docker.operators.docker_swarm import DockerSwarmOperator
-
-default_args = {
-    'owner': 'airflow',
-    'depends_on_past': False,
-    'start_date': days_ago(1),
-    'email': ['airflow@example.com'],
-    'email_on_failure': False,
-    'email_on_retry': False
-}
-
-dag = DAG(
-    'docker_swarm_sample',
-    default_args=default_args,
-    schedule_interval=timedelta(minutes=10),
-    catchup=False
-)
-
-with dag as dag:
-    t1 = DockerSwarmOperator(
-        api_version='auto',
-        docker_url='tcp://localhost:2375', # Set your docker URL
-        command='/bin/sleep 10',
-        image='centos:latest',
-        auto_remove=True,
-        task_id='sleep_with_swarm',
-    )
-"""
